WIP: Analyze LogQL pipelines

+++ pkg/chunkenc/unordered.go
@@ -275,7 +275,7 @@ func (hb *unorderedHeadBlock) Iterator(
 	// todo(we should record the real duration and the actual samples out from the last child
 	// in should also be the real amount of entries returned by the original query to the ingester
 	in, _ := a.GetChild(0).GetCounts()
-	a.Set(time.Second, in, totalSamples)
+	a.Set(time.Second, in, int64(totalSamples))
 
 	return iter.NewStreamsIterator(streamsResult, direction)
 }
